Add files via upload
this is with a nested dictionary...

not sure if you wanna use this so then it might be easier to access the keys (i.e. comment_text) to further analyse without removing the order i.e. index, label and comment_text all together...

otherwise maybe we can do this with a list of lists // list/array of tuples but then not sure - can we still maintain the row order and also analyse the text?????